[All] Code style
