+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamo/impl/database/statementhandler/SqlDynamicDtObject.java
@@ -23,7 +23,6 @@ import io.vertigo.dynamo.domain.metamodel.DtField;
 import io.vertigo.dynamo.domain.metamodel.Dynamic;
 import io.vertigo.dynamo.domain.model.DtObject;
 import io.vertigo.dynamo.domain.util.DtObjectUtil;
-import io.vertigo.dynamo.impl.database.statementhandler.SqlResultMetaDataDynamic.SerializableDtDefinition;
 import io.vertigo.lang.Assertion;
 
 import java.util.HashMap;
